Update README.md for Mac Building (#8)
The libtool requirement was missing in the MacOS instructions. Added that to both the Homebrew and MacPorts sections. Additionally, condensed the Homebrew install section to a one-liner, since that works great and is in keeping with the `pacman` instructions for Arch. I don't know how MacPorts works so I left that as multi-line.